Generate Tamogotchi Code
This code changes the scanner input to a static so that it can be more easily used in GenerateTamogotchi. This is so I did not have to generate another scanner that can not be closed due to closing System.In if a scanner is closed.

This code has a method for each question, for more easy editing, adding, and removing questions.
This will use the sum value of answers to determine the Tamo model to use.